Merge main with booking

*Prepare for general e2e test

+++ src/components/organisms/Clients.tsx
@@ -2,19 +2,22 @@ import Button from 'components/atoms/Button';
 import Header from 'components/atoms/Header';
 import MultipleRecords from 'components/atoms/MultipleRecords';
 import NewClientForm from 'components/molecules/forms/NewClientForm';
-import { IAdminState, IClient, IReduxState } from 'models';
+import { IBooking, IClient, instanceOfClients, IReduxState } from 'models';
 import * as React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { deleteClient, openModal } from 'store';
+import { closeModal, deleteClient, openModal } from 'store';
 import styled from 'styled-components';
 import {
   MODAL_TYPES,
+  RECORDS_CLIENTS_DETAILS_PROPERTY_MAP,
   RECORDS_CLIENTS_HEADERS,
   RECORDS_CLIENTS_ROW,
   RECORDS_CLIENTS_ROW_DETAILS
 } from 'utils';
 import Modal from 'components/organisms/Modal';
 import SearchInputField from 'components/atoms/SearchInputField';
+import ModalDelete from 'components/molecules/modals/ModalDelete';
+import ModalInfo from 'components/molecules/modals/ModalInfo';
 
 const CalenderWrapper = styled.section`
   width: 60%;
@@ -34,26 +37,24 @@ const RecordsActionContent = styled.div`
   align-items: center;
   justify-content: space-between;
 `;
+
 const OpenClientModalButton = styled(Button)`
   background-color: #eaeaea;
-  border-color: #afbf36;
+  border-color: ${({ theme }) => theme.green};
   color: #454545;
   &:hover {
-    background-color: #afbf36;
+    background-color: ${({ theme }) => theme.green};
     border-color: #b9b8b8;
     box-shadow: none;
     opacity: 1;
   }
 `;
 
-interface IProps {
-  mainState: IAdminState;
-}
-
-const Clients: React.FunctionComponent<IProps> = ({ mainState }) => {
+const Clients = () => {
   const [clientList, setClientList] = React.useState<IClient[]>([]);
   const [isEditing, setIsEditing] = React.useState(false);
   const [editedItemIndex, setEditedItemIndex] = React.useState<number | undefined>(undefined);
+  const [deleteItemIndex, setDeleteItemIndex] = React.useState<number | undefined>(undefined);
 
   const dispatch = useDispatch();
 
@@ -62,28 +63,44 @@ const Clients: React.FunctionComponent<IProps> = ({ mainState }) => {
     modal: { isOpen, type }
   } = useSelector((state: IReduxState) => state);
 
-  const editClientHandler = (index: number) => {
+  const editClientHandler = (index: number, editHandler: boolean) => {
     setIsEditing(true);
     setEditedItemIndex(index);
     dispatch(openModal(MODAL_TYPES.CLIENT));
   };
 
   const deleteClientHandler = (index: number) => {
-    const currentClient = clients[index];
-    if (currentClient.id) dispatch(deleteClient(currentClient.id));
-    initialEditingState();
+    setDeleteItemIndex(index);
+    dispatch(openModal(MODAL_TYPES.DELETE));
   };
 
-  const initialEditingState = () => {
+  const deleteClientAction = () => {
+    if (typeof deleteItemIndex === 'undefined') return;
+    const currentClient = clients[deleteItemIndex];
+    if (currentClient.id) dispatch(deleteClient(currentClient.id));
+    clientListHandler(clients.filter((c) => c.id !== currentClient.id));
+    initialClientState();
+    dispatch(closeModal());
+  };
+
+  const cancelDeleteClientAction = () => {
+    initialClientState();
+    dispatch(closeModal());
+  };
+
+  const initialClientState = () => {
     setIsEditing(false);
     setEditedItemIndex(undefined);
+    setDeleteItemIndex(undefined);
   };
 
-  const clientListHandler = (searchResults: IClient[]): void => {
-    setClientList(searchResults);
+  const clientListHandler = (searchResults: (IClient | IBooking)[]): void => {
+    if (searchResults.length && instanceOfClients(searchResults)) {
+      setClientList(searchResults);
+    }
   };
 
-  React.useEffect(() => undefined, [clientList]);
+  React.useEffect(() => setClientList(clients), [clients]);
 
   return (
     <CalenderWrapper>
@@ -93,6 +110,7 @@ const Clients: React.FunctionComponent<IProps> = ({ mainState }) => {
           type="clients"
           placeholder="Wyszukaj najemce"
           searchContent={clients}
+          searchProperty="name"
           searchContentHandler={clientListHandler}
         />
         <OpenClientModalButton onClick={() => dispatch(openModal(MODAL_TYPES.CLIENT))}>
@@ -100,9 +118,11 @@ const Clients: React.FunctionComponent<IProps> = ({ mainState }) => {
         </OpenClientModalButton>
       </RecordsActionContent>
       <MultipleRecords
+        isAdmin
         headers={RECORDS_CLIENTS_HEADERS}
         dataProperty={RECORDS_CLIENTS_ROW}
         dataPropertyDetails={RECORDS_CLIENTS_ROW_DETAILS}
+        dataPropertyDisplayMap={RECORDS_CLIENTS_DETAILS_PROPERTY_MAP}
         records={clientList}
         editHandler={editClientHandler}
         deleteHandler={deleteClientHandler}
@@ -114,9 +134,17 @@ const Clients: React.FunctionComponent<IProps> = ({ mainState }) => {
             <NewClientForm
               isEditing={isEditing}
               editedItemIndex={editedItemIndex}
-              initialEditingState={initialEditingState}
+              initialEditingState={initialClientState}
             />
           )}
+          {type === MODAL_TYPES.DELETE && (
+            <ModalDelete
+              message="Czy na pewno chcesz chcesz skazować tego clienta"
+              callback={deleteClientAction}
+              cancelCallback={cancelDeleteClientAction}
+            />
+          )}
+          {type === MODAL_TYPES.SUCCESS && <ModalInfo header="Klienci" />}
         </Modal>
       )}
     </CalenderWrapper>
